Added Args for Reverse and for Specifying a File
Added the Following Command
--reverse
Which causes the output to be reversed
Also Added the ability to pick any file instead of just SortMe.txt

Fixes:
I forgot how arrays worked, and had it checking the 2nd and 3rd character in a word. This lead to the list not being sorted correctly. The program now correctly checks the 1st and 2nd letter.