brokeman act 1
Narrator: No one was left who could remember how it had happened,
how the world had fallen under darkness.
At least no one who would do anything.
No one who would oppose the robots.
No one who would challenge their power,
or so Dr. Wily believed...

Twenty floors above the dark streets of the city,
Dr. Light lived in a run-down tenement.
An eccentric and brilliant man.
Light was a loner, a thinker, a man of ideas.
Ideas forbidden in Wily's society.
The society for which he worked.
The society in which he lived.
The society that he would set free.
And so Light worked, far into the night,
when the watchful eyes of Wily's robots weren't upon him.
He'd set his skillful hands to the task of creating a device
to bring about a change, to create a machine to bring freedom,
to create a man to save the world.
Twelve years Light worked and on a cold night in the year 200X,
Protoman was born.
A perfect man, an unbeatable machine,
hell-bent on destroying every evil standing between man and freedom,
built for one purpose, to destroy Wily's army of evil robots.
Ready, willing, prepared to fight.

Cutman
Gutsman
Elecman
Bombman
Fireman
Iceman
Proto

Fireman: Attack!

Narrator: And as the smoke cleared!
Wily rose above the countless robots remaining.
Protoman was wounded, low on energy,
struggling to remain standing as Wily ordered the final attack.
The death of Protoman.

The crowd had gathered there to watch him fall,
to watch their hopes destroyed.
They watched them beat him, they watched them break him,
they watched his last defense deployed.
There was not a man among them who would let himself be heard.
But from the crowd, from their collective fear,
arose these broken words:
We are the dead
We are the dead

Human Choir: What have we done?
Narrator: We are the dead
Human Choir: What will we do?
Narrator: We are the dead
Human Choir: Where will we turn?
Narrator: We are the dead
Human Choir: Is there nothing we can do?
Narrator: We are the dead
Human Choir: How did it come to this?
Narrator: We are the dead
Human Choir: How did we go so wrong?
Narrator: We are the dead
Human Choir: We are the dead