brokeman act 2
this fixes the bug

Albert:
Tonight the streets are red,
the lights are blue and blinding.
No sign of the "Good Doctor,"
but the siren's wail and whining
Tell us he'll be found.

I can almost hear the hounds...

What kind of man builds a machine to kill a girl?
No he did not use his hands.
Like a smart man he used a tool.
But just the same,
How can you question who's to blame?

Reporters:
What was her name?

Albert:
It doesn't matter.
Now listen...
"The Good Doctor" has to pay!

When I say he was a monster,
when I set fire to his name,
It does not matter where you hear it from,
whether truth or lies,
It gets said all the same.

Whatever's on the table plays!